build: Bash completion
https://phabricator.endlessm.com/T20365

Running `setup.py build_completion` a bash script will be created
inside the dist folder.  This script can be sourced into a bash
session to get completion for the command flapjack.

Implementation: there is a template for the bash script which is
filled with information obtained from the parser of current
subcommands.  This is a bit hacky.

Future improvement: the script should be installed in the system, the
places seem to be:
- /etc/bash_completion.d/
- /usr/share/bash-completion/completions/

And for user: ~/.bash_completion (but is a file that can exist, not a
directory)